Removed baseURL from package.json
This is needed in order for jspm NOT to generate a baseURL into the system.config.js.

The Problem with this value is related to karma-jspm. which trys to override it during the test-run. Therefore it is not specified in this config anymore, but during the build process or test-process only.