Added editmode for labeledthings in frame
There are still a lot of bugs in this thing
Furthermore the datastructure is far from ideal.

I still hope it will work for the presentation tommorow.

Selections are shown only for a second, after disappearing again, the context on the right is still assigned with the selection and can edit it. This needs further investigation. Eventhough I have some ideas about why this is occuring.