puppet: install g++ before aglio
The reason that aglio always failed was because it is compiled using
`g++` which will be installed as a dependency for `php-dev` but `aglio`
is tried to be installed first.
However, it makes no sense to to make aglio depend on php-dev, so
I added this dependency directly.
Now, the `vagrant up` again runs without any problems, at least for me.
;-)