labeling_api: refactor to automatically generate missing uuids in database facades
This patch also includes some other refactorings which I did while
I read the source because I couldn't leave the code as it was and the
changes got too big to revert and commit them separately. I am sorry for
that but I think the world is a better place now. ;-)