labeling_api: rename all `labelingTaskId` properties to `taskId`
The UI uses `taskId` and so should we.

However, this patch currently only changes this property in application
code but leaves the model properties untouched and instead uses the jms
serializer to rename those properties during serialization.

So, this change should be backwards compatible.

However, we should rename the model properties, too, sometime and remove
the serializer annotations.