labeling_api: introduce classes for shapes
This puts almost any 'real' logic of the KITTI exporter into the shape
classes and makes it really easy to work with the shapes.
So, any shape de-/serialization steps are concentrated at on place (the
shape classes), which eliminates the need of `createRectangleShape` and
similar methods in the tests.