labeling_api: remove setter LabeledThingInFrame::setLabeledThingId
Using reflection here allows us to enforce a valid `LabeledThing` as
constructor parameter and we no longer require a setter for the
labeledThingId. This is important because an upcoming change
additionally adds the taskId to the LabeledThingInFrame model which is
intended to speed up some things and we don't want to have the
possibility to set the labeledThingId without a taskId or an invalid taskId.
So, the idea here is to make the code a) more readable and b) more
robust.