Fixed a bug where no bracket could be placed left of the current thumbnail
angular-dragdrop already calls $eval on the data-drop argument which
in conjunction with the interpolation braces leads to the function passed
being converted to a constant boolean. This again leads to angular-dragdrop
cancelling its watch on data-drop for performance reasons, thus not updating
the disabled state of drop targets after the initial page load.