labeling-api: refactor api documentation
I found the feature defining data structures upfront which then can be
referred to later on in a very convinient way. By using this, arrays are
always represented by just one element for example but we only have one
type definition and this type is always the same throughout the whole
documentation which is very nice, imo.