Disabled all PouchDb related tests for now
This is done to check piece by piece, which of those tests are matching the needed feature set and are implemented in a working way.

Currently most of them simply fail due to a test<->implementation mismatch, or due to wrongly done test bootstrapping.

They seem not have been executed in a long time. (Or maybe never?)