Fixed TimerGateway structure
During the pouchdb implementation an incompatibility between the used timer structure on the server side and the frontend side was introduced. I would assume this was done in order to quickly test an implementation idea and then wrongfully comitted. This commit changes the corresponding interface back to its original design.