Added migration to fix inline `labeledThingInFrameId` entries
The `ghostBust` function of `LabeledThingInFrame` models did correctly update the `id` of the LTIF itself, but not the `labeledThingInFrameId` reference inside its own shape array. Therefore every manifested ghost has most likely the wrong id stored in the shapes array.

This migration script updates the id of each shape to the if of the LTIF it is attached to.